Update GameBoard.py
Pieces=-1 is now set as a default for the constructor of GameBoard, so if you want a board with a random number of pieces, just call GameBoard() with no arguments.

Added all_true(arr) as a helper function for desired_probabilities_2, which is used to check if the "finished" array is entirely true, meaning all games are finished.

Updated desired_probabilities_2, which now ends early if all games are finished, and plays an entire game up to 42 pieces. Slightly changed some variable names for more clarity, and tweaked the way finished games are tracked.