Create SimpleNN.py
This is a simple from-scratch neural network that I created in early May (initially in Java and then translated to Python). The current version only has an input layer and output layer, and the task it completes is very simple (for a sequence of zeros and ones, it outputs 0 if the first and last digit are the same, 1 if the first is 1, and -1 if the last is 1). It is certainly not very useful on its own, but it may be able to serve as a rough starting point for a from-scratch network that allows for more dynamic architecture (I will attempt to experiment with this in the near future when I get the chance).