Create MirroredStrategy
An alternate version of the neural network code altered to make use of a mirrored strategy to allow running on multiple GPUs. Changes are made to the format of the training data to disable "auto sharding", which causes errors when using mirrored strategy. Testing is required to ensure that these alterations have not damaged the learning capabilities of the network.